Pin svglib to a version below 1.6.0 (#1550) (#1551)
* fix(deps): Pin svglib to 1.5.1 due to pycairo installation failure (#1550)

The latest svglib release (1.6.0) has introduced a new dependency on pycairo, which requires system-level libraries to install correctly. This change causes the installation of easy-thumbnails[svg] to fail in environments where these libraries are not present.

This commit pins svglib to version 1.5.1 as a temporary workaround until the upstream issue is resolved.

Ref: https://github.com/deeplook/svglib/issues/421

* fix (build): revert flake8 version update because of incampatability with pyproject-flake8